Ayush | refactor: refactored some code and fixed test cases
